input validation added

+++ routers/roomsRouter/roomsRouter.js
@@ -1,34 +1,37 @@
 import { Router } from "express";
-import db from "../../database/database.js";
 import { authenticateToken } from "../middleware/verifyJWT.js";
 import { logErrorToFile } from "../../utils/logErrorToFile/logErrorToFile.js";
 import { deleteRateLimiter, generalRateLimiter, postRateLimiter } from "../middleware/rateLimit.js";
+import { validateInput } from "../../utils/inputValidation/inputValidation.js";
+import db from "../../database/database.js";
+
 const router = Router();
 
 router.get("/api/rooms/:institutionid", authenticateToken, generalRateLimiter, async (req, res) => {
-  const institutionId = req.params.institutionid;
-  const current_role_name = req.user.role_name
-  if (current_role_name != "admin"){
-    res.status(403).send({message: 'Forbidden'})
-  } 
   try {
-  const query = `
-  SELECT 
-      ro.id AS room_id,
-      ro.room_name,
-      ro.institution_id,
-      GROUP_CONCAT(JSON_OBJECT('id', c.id, 'course_name', c.course_name)) AS courses
-  FROM 
-      rooms ro
-  LEFT JOIN 
-      rooms_courses rc ON ro.id = rc.room_id
-  LEFT JOIN 
-      courses c ON rc.course_id = c.id
-  WHERE 
-      ro.institution_id = ?
-  GROUP BY 
-      ro.id;
-  `;
+  
+    const institutionId = req.params.institutionid;
+    const current_role_name = req.user.role_name
+    if (current_role_name != "admin"){
+      res.status(403).send({message: 'Forbidden'})
+    } 
+    const query = `
+    SELECT 
+        ro.id AS room_id,
+        ro.room_name,
+        ro.institution_id,
+        GROUP_CONCAT(JSON_OBJECT('id', c.id, 'course_name', c.course_name)) AS courses
+    FROM 
+        rooms ro
+    LEFT JOIN 
+        rooms_courses rc ON ro.id = rc.room_id
+    LEFT JOIN 
+        courses c ON rc.course_id = c.id
+    WHERE 
+        ro.institution_id = ?
+    GROUP BY 
+        ro.id;
+    `;
 
     const [results] = await db.connection.query(query, [institutionId]);
 
@@ -49,17 +52,23 @@ router.get("/api/rooms/:institutionid", authenticateToken, generalRateLimiter, a
 
 //Post a single room
 router.post("/api/rooms", authenticateToken, postRateLimiter, async (req, res) => {
-  const body = req.body;
-
-  const current_role_name = req.user.role_name
-  if (current_role_name != "admin"){
-    res.status(403).send({message: 'Forbidden'})
-  } 
-  if (!body.roomName && !body.institutionId) {
-    return res.status(400).send({ message: "Bad Reqeust" });
-  }
-
   try {
+    const body = req.body;
+
+    const current_role_name = req.user.role_name
+    if (current_role_name != "admin"){
+      res.status(403).send({message: 'Forbidden'})
+    }
+
+    const validation = await validateInput(req.body);
+    if (!validation) {
+      return res.status(400).json({ message: "Bad Request" });
+    }
+
+    if (!body.roomName && !body.institutionId) {
+      return res.status(400).send({ message: "Bad Request" });
+    }
+
     const [existingRoom] = await db.connection.query("SELECT * FROM rooms WHERE room_name = ? AND institution_id = ?",[body.roomName, body.institutionId]);
 
     if (existingRoom.length > 0) {
@@ -77,20 +86,24 @@ router.post("/api/rooms", authenticateToken, postRateLimiter, async (req, res) =
 });
 
 //Post a room and a course into the reference table rooms_courses
-
 router.post("/api/rooms/courses", authenticateToken, postRateLimiter, async (req, res) => {
-
-  const current_role_name = req.user.role_name
-  if (current_role_name != "admin"){
-    res.status(403).send({message: 'Forbidden'})
-  } 
-  const roomsCourses = req.body.assigned;
-  if (!roomsCourses || roomsCourses.length === 0) {
-    return res.status(400).send({ message: "Bad Request" });
-  }
   try {
-  const values = roomsCourses.map(({ roomId, courseId }) => [roomId, courseId]);
-  const query = `INSERT INTO rooms_courses (room_id, course_id) VALUES ?`;
+    const current_role_name = req.user.role_name
+    if (current_role_name != "admin"){
+      res.status(403).send({message: 'Forbidden'})
+    }
+
+    const validation = await validateInput(req.body);
+    if (!validation) {
+      return res.status(400).json({ message: "Bad Request" });
+    }
+
+    const roomsCourses = req.body.assigned;
+    if (!roomsCourses || roomsCourses.length === 0) {
+      return res.status(400).send({ message: "Bad Request" });
+    }
+    const values = roomsCourses.map(({ roomId, courseId }) => [roomId, courseId]);
+    const query = `INSERT INTO rooms_courses (room_id, course_id) VALUES ?`;
     await db.connection.query(query, [values]);
     res.send({message:`Successfully assigned course to room`, assigned: roomsCourses});
   } catch (error) {
@@ -100,22 +113,28 @@ router.post("/api/rooms/courses", authenticateToken, postRateLimiter, async (req
 });
 
 router.delete("/api/rooms/courses", authenticateToken, deleteRateLimiter, async (req, res) => {
-  const current_role_name = req.user.role_name
-  if (current_role_name != "admin"){
-    res.status(403).send({message: 'Forbidden'})
-  } 
-  const roomsCourses = req.body.removed;
-  if (!roomsCourses || roomsCourses.length === 0) {
-    return res.status(400).send({ message: "Bad Request"});
-  }
-
   try {
-  const query = `
-      DELETE FROM rooms_courses
-      WHERE (room_id, course_id) IN (?)
-  `;
+    const current_role_name = req.user.role_name
+    if (current_role_name != "admin"){
+      res.status(403).send({message: 'Forbidden'})
+    } 
 
-  const values = roomsCourses.map(({ roomId, courseId }) => [roomId, courseId]);
+    const validation = await validateInput(req.body);
+    if (!validation) {
+      return res.status(400).json({ message: "Bad Request" });
+    }
+
+    const roomsCourses = req.body.removed;
+    if (!roomsCourses) {
+      return res.status(400).send({ message: "Bad Request"});
+    }
+
+    const query = `
+        DELETE FROM rooms_courses
+        WHERE (room_id, course_id) IN (?)
+    `;
+
+    const values = roomsCourses.map(({ roomId, courseId }) => [roomId, courseId]);
 
     await db.connection.query(query, [values]);
     res.send({message: `Successfully removed course from room`, deleted: roomsCourses });
@@ -126,12 +145,18 @@ router.delete("/api/rooms/courses", authenticateToken, deleteRateLimiter, async 
 });
 
 router.delete("/api/rooms/:roomid", authenticateToken, deleteRateLimiter, async (req, res) => {
-  const roomId = req.params.roomid;
   try {
+    const roomId = req.params.roomid;
     const current_role_name = req.user.role_name
     if (current_role_name != "admin"){
       res.status(403).send({message: 'Forbidden'})
-    } 
+    }
+
+    const validation = await validateInput(req.body);
+    if (!validation) {
+      return res.status(400).json({ message: "Bad Request" });
+    }
+
     const [result] = await db.connection.query("DELETE FROM rooms WHERE id = ?", [roomId]);
     res.send({message: `Successfully deleted room`, data: result});
   } catch (error) {
